infra: extract listing tests to separate files
- `tests` -> `src/tests/mod.rs`
- `tests::config` -> `src/tests/config.rs`

This is just profoundly easier to work with.
