+++ packages/mdbook-trpl-listing/src/lib.rs
@@ -361,387 +361,4 @@ impl<'a> ListingBuilder<'a> {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    /// Note: This inserts an additional backtick around the re-emitted code.
-    /// It is not clear *why*, but that seems to be an artifact of the rendering
-    /// done by the `pulldown_cmark_to_cmark` crate.
-    #[test]
-    fn default_mode_works() {
-        let result = rewrite_listing(
-            r#"<Listing number="1-2" caption="A write-up which *might* include inline Markdown like `code` etc." file-name="src/main.rs">
-
-```rust
-fn main() {}
-```
-
-</Listing>"#,
-            Mode::Default,
-        );
-
-        assert_eq!(
-            &result.unwrap(),
-            r#"<figure class="listing">
-<span class="file-name">Filename: src/main.rs</span>
-
-````rust
-fn main() {}
-````
-
-<figcaption>Listing 1-2: A write-up which <em>might</em> include inline Markdown like <code>code</code> etc.</figcaption>
-</figure>"#
-        );
-    }
-
-    #[test]
-    fn simple_mode_works() {
-        let result = rewrite_listing(
-            r#"<Listing number="1-2" caption="A write-up which *might* include inline Markdown like `code` etc." file-name="src/main.rs">
-
-```rust
-fn main() {}
-```
-
-</Listing>"#,
-            Mode::Simple,
-        );
-
-        assert_eq!(
-            &result.unwrap(),
-            r#"
-Filename: src/main.rs
-
-````rust
-fn main() {}
-````
-
-Listing 1-2: A write-up which <em>might</em> include inline Markdown like <code>code</code> etc."#
-        );
-    }
-
-    #[test]
-    fn actual_listing() {
-        let result = rewrite_listing(
-            r#"Now open the *main.rs* file you just created and enter the code in Listing 1-1.
-
-<Listing number="1-1" file-name="main.rs" caption="A program that prints `Hello, world!`">
-
-```rust
-fn main() {
-    println!("Hello, world!");
-}
-```
-
-</Listing>
-
-Save the file and go back to your terminal window"#,
-            Mode::Default,
-        );
-
-        assert!(result.is_ok());
-        assert_eq!(
-            result.unwrap(),
-            r#"Now open the *main.rs* file you just created and enter the code in Listing 1-1.
-
-<figure class="listing">
-<span class="file-name">Filename: main.rs</span>
-
-````rust
-fn main() {
-    println!("Hello, world!");
-}
-````
-
-<figcaption>Listing 1-1: A program that prints <code>Hello, world!</code></figcaption>
-</figure>
-
-Save the file and go back to your terminal window"#
-        );
-    }
-
-    #[test]
-    fn no_filename() {
-        let result = rewrite_listing(
-            r#"This is the opening.
-
-<Listing number="1-1" caption="This is the caption">
-
-```rust
-fn main() {}
-```
-
-</Listing>
-
-This is the closing."#,
-            Mode::Default,
-        );
-
-        assert!(result.is_ok());
-        assert_eq!(
-            result.unwrap(),
-            r#"This is the opening.
-
-<figure class="listing">
-
-````rust
-fn main() {}
-````
-
-<figcaption>Listing 1-1: This is the caption</figcaption>
-</figure>
-
-This is the closing."#
-        );
-    }
-
-    /// Check that the config options are correctly handled.
-    ///
-    /// Note: none of these tests particularly exercise the *wiring*. They just
-    /// assume that the config itself is done correctly. This is a small enough
-    /// chunk of code that it easy to verify by hand at present. If it becomes
-    /// more complex in the future, it would be good to revisit and integrate
-    /// the same kinds of tests as the unit tests above here.
-    #[cfg(test)]
-    mod config {
-        use super::*;
-
-        // TODO: what *should* the behavior here be? I *think* it should error,
-        // in that there is a problem if it is invoked without that info.
-        #[test]
-        fn no_config() {
-            let input_json = r##"[
-                {
-                    "root": "/path/to/book",
-                    "config": {
-                        "book": {
-                            "authors": ["AUTHOR"],
-                            "language": "en",
-                            "multilingual": false,
-                            "src": "src",
-                            "title": "TITLE"
-                        },
-                        "preprocessor": {}
-                    },
-                    "renderer": "html",
-                    "mdbook_version": "0.4.21"
-                },
-                {
-                    "sections": [
-                        {
-                            "Chapter": {
-                                "name": "Chapter 1",
-                                "content": "# Chapter 1\n",
-                                "number": [1],
-                                "sub_items": [],
-                                "path": "chapter_1.md",
-                                "source_path": "chapter_1.md",
-                                "parent_names": []
-                            }
-                        }
-                    ],
-                    "__non_exhaustive": null
-                }
-            ]"##;
-            let input_json = input_json.as_bytes();
-            let (ctx, book) =
-                mdbook::preprocess::CmdPreprocessor::parse_input(input_json)
-                    .unwrap();
-            let result = TrplListing.run(&ctx, book);
-            assert!(result.is_err());
-            let err = result.unwrap_err();
-            assert_eq!(format!("{err}"), "No config for trpl-listing");
-        }
-
-        #[test]
-        fn empty_config() {
-            let input_json = r##"[
-                {
-                    "root": "/path/to/book",
-                    "config": {
-                        "book": {
-                            "authors": ["AUTHOR"],
-                            "language": "en",
-                            "multilingual": false,
-                            "src": "src",
-                            "title": "TITLE"
-                        },
-                        "preprocessor": {
-                            "trpl-listing": {}
-                        }
-                    },
-                    "renderer": "html",
-                    "mdbook_version": "0.4.21"
-                },
-                {
-                    "sections": [
-                        {
-                            "Chapter": {
-                                "name": "Chapter 1",
-                                "content": "# Chapter 1\n",
-                                "number": [1],
-                                "sub_items": [],
-                                "path": "chapter_1.md",
-                                "source_path": "chapter_1.md",
-                                "parent_names": []
-                            }
-                        }
-                    ],
-                    "__non_exhaustive": null
-                }
-            ]"##;
-            let input_json = input_json.as_bytes();
-            let (ctx, book) =
-                mdbook::preprocess::CmdPreprocessor::parse_input(input_json)
-                    .unwrap();
-            let result = TrplListing.run(&ctx, book);
-            assert!(result.is_ok());
-        }
-
-        #[test]
-        fn specify_default() {
-            let input_json = r##"[
-                {
-                    "root": "/path/to/book",
-                    "config": {
-                        "book": {
-                            "authors": ["AUTHOR"],
-                            "language": "en",
-                            "multilingual": false,
-                            "src": "src",
-                            "title": "TITLE"
-                        },
-                        "preprocessor": {
-                            "trpl-listing": {
-                                "output-mode": "default"
-                            }
-                        }
-                    },
-                    "renderer": "html",
-                    "mdbook_version": "0.4.21"
-                },
-                {
-                    "sections": [
-                        {
-                            "Chapter": {
-                                "name": "Chapter 1",
-                                "content": "# Chapter 1\n",
-                                "number": [1],
-                                "sub_items": [],
-                                "path": "chapter_1.md",
-                                "source_path": "chapter_1.md",
-                                "parent_names": []
-                            }
-                        }
-                    ],
-                    "__non_exhaustive": null
-                }
-            ]"##;
-            let input_json = input_json.as_bytes();
-            let (ctx, book) =
-                mdbook::preprocess::CmdPreprocessor::parse_input(input_json)
-                    .unwrap();
-            let result = TrplListing.run(&ctx, book);
-            assert!(result.is_ok());
-        }
-
-        #[test]
-        fn specify_simple() {
-            let input_json = r##"[
-                {
-                    "root": "/path/to/book",
-                    "config": {
-                        "book": {
-                            "authors": ["AUTHOR"],
-                            "language": "en",
-                            "multilingual": false,
-                            "src": "src",
-                            "title": "TITLE"
-                        },
-                        "preprocessor": {
-                            "trpl-listing": {
-                                "output-mode": "simple"
-                            }
-                        }
-                    },
-                    "renderer": "html",
-                    "mdbook_version": "0.4.21"
-                },
-                {
-                    "sections": [
-                        {
-                            "Chapter": {
-                                "name": "Chapter 1",
-                                "content": "# Chapter 1\n",
-                                "number": [1],
-                                "sub_items": [],
-                                "path": "chapter_1.md",
-                                "source_path": "chapter_1.md",
-                                "parent_names": []
-                            }
-                        }
-                    ],
-                    "__non_exhaustive": null
-                }
-            ]"##;
-            let input_json = input_json.as_bytes();
-            let (ctx, book) =
-                mdbook::preprocess::CmdPreprocessor::parse_input(input_json)
-                    .unwrap();
-            let result = TrplListing.run(&ctx, book);
-            assert!(result.is_ok());
-        }
-
-        #[test]
-        fn specify_invalid() {
-            let input_json = r##"[
-                {
-                    "root": "/path/to/book",
-                    "config": {
-                        "book": {
-                            "authors": ["AUTHOR"],
-                            "language": "en",
-                            "multilingual": false,
-                            "src": "src",
-                            "title": "TITLE"
-                        },
-                        "preprocessor": {
-                            "trpl-listing": {
-                                "output-mode": "nonsense"
-                            }
-                        }
-                    },
-                    "renderer": "html",
-                    "mdbook_version": "0.4.21"
-                },
-                {
-                    "sections": [
-                        {
-                            "Chapter": {
-                                "name": "Chapter 1",
-                                "content": "# Chapter 1\n",
-                                "number": [1],
-                                "sub_items": [],
-                                "path": "chapter_1.md",
-                                "source_path": "chapter_1.md",
-                                "parent_names": []
-                            }
-                        }
-                    ],
-                    "__non_exhaustive": null
-                }
-            ]"##;
-            let input_json = input_json.as_bytes();
-            let (ctx, book) =
-                mdbook::preprocess::CmdPreprocessor::parse_input(input_json)
-                    .unwrap();
-            let result = TrplListing.run(&ctx, book);
-            assert!(result.is_err());
-            let err = result.unwrap_err();
-            assert_eq!(
-                format!("{err}"),
-                "Bad config value '\"nonsense\"' for key 'output-mode'"
-            );
-        }
-    }
-}
+mod tests;
